# services/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/ai/chat": {
      "post": {
        "operationId": "AiGatewayController_chat",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Chat completion response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_ChatCompletionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a chat completion request",
        "tags": [
          "AI Gateway"
        ]
      }
    },
    "/api/v1/ai/chat/stream": {
      "post": {
        "operationId": "AiGatewayController_chatStream",
        "parameters": [],
        "responses": {
          "200": {
            "description": "SSE stream of chat completion chunks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stream a chat completion response via SSE",
        "tags": [
          "AI Gateway"
        ]
      }
    },
...
    "/api/v1/knowledge/documents": {
      "post": {
        "operationId": "DocumentController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Document uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a document",
        "tags": [
          "Knowledge Documents"
        ]
      },
      "get": {
        "operationId": "DocumentController_findMany",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List documents",
        "tags": [
          "Knowledge Documents"
        ]
      }
    },
    "/api/v1/knowledge/documents/{id}": {
      "get": {
        "operationId": "DocumentController_findById",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Document found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get document by ID",
        "tags": [
          "Knowledge Documents"
        ]
      },
      "patch": {
        "operationId": "DocumentController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update document metadata",
        "tags": [
          "Knowledge Documents"
        ]
      },
      "delete": {
        "operationId": "DocumentController_delete",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Document deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a document",
        "tags": [
          "Knowledge Documents"
        ]
      }
    },
    "/api/v1/knowledge/documents/{id}/archive": {
      "post": {
        "operationId": "DocumentController_archive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a document",
        "tags": [
          "Knowledge Documents"
        ]
      }
    },
    "/api/v1/knowledge/documents/{id}/restore": {
      "post": {
        "operationId": "DocumentController_restore",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore a deleted document",
        "tags": [
          "Knowledge Documents"
        ]
      }
    },
    "/api/v1/knowledge/documents/{id}/download": {
      "get": {
        "operationId": "DocumentController_download",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download a document",
        "tags": [
          "Knowledge Documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "Atlas AI API",
    "description": "Enterprise AI Platform API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TokenUsageDto": {
        "type": "object",
        "properties": {
          "promptTokens": {
            "type": "number",
            "description": "Number of prompt tokens",
            "example": 50
          },
          "completionTokens": {
            "type": "number",
            "description": "Number of completion tokens",
            "example": 100
          },
          "totalTokens": {
            "type": "number",
            "description": "Total tokens used",
            "example": 150
          },
          "estimatedCost": {
            "type": "number",
            "description": "Estimated cost in USD",
            "example": 0.002
          }
        },
        "required": [
          "promptTokens",
          "completionTokens",
          "totalTokens",
          "estimatedCost"
        ]
      },
      "_ChatCompletionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Response ID"
          },
          "model": {
            "type": "string",
            "description": "Model used"
          },
          "provider": {
            "type": "string",
            "description": "Provider name"
          },
          "content": {
            "type": "string",
            "description": "Generated content"
          },
          "finishReason": {
            "type": "string",
            "enum": [
              "stop",
              "length",
              "error"
            ],
            "description": "Finish reason"
          },
          "usage": {
            "description": "Token usage",
            "allOf": [
              {
                "$ref": "#/components/schemas/TokenUsageDto"
              }
            ]
          },
          "latency": {
            "type": "number",
            "description": "Latency in ms"
          }
        },
        "required": [
          "id",
          "model",
          "provider",
          "content",
          "finishReason",
          "usage",
          "latency"
        ]
      },
...
    }
  }
}